Updated code for directory index file generation

+++ python/datamodel/directory_index_generate.py
@@ -11,6 +11,7 @@ from astropy.io import fits
 from astropy.table import Table
 from jinja2 import Environment, FileSystemLoader
 from collections import deque
+import pdb
 
 # example environment path to some data
 os.environ['DR'] = (pathlib.Path('../').resolve()).as_posix()
@@ -26,12 +27,14 @@ class directory_index_generator(object):
     dir_deque = deque()
     
     def __init__(self,
-            base_html_path = f'{dr_base_path}/data_model/docs',
+            base_html_path = f'{dr_base_path}/data_model/public-html',
             base_yaml_path = f'{dr_base_path}/data_model/products/yaml',
         ):
         
         self.base_html_path = base_html_path
         self.base_yaml_path = base_yaml_path
+        
+        self.release = release_version
         
         self.setup_jinja2()
         
@@ -40,7 +43,9 @@ class directory_index_generator(object):
     def setup_jinja2(self):
         """ Set up the Jinja2 template enviroment """
         loader = FileSystemLoader("templates")
-        self.environment = Environment(loader=loader, trim_blocks=True, lstrip_blocks=True)
+        self.environment = Environment(
+            loader=loader, trim_blocks=True, lstrip_blocks=True,
+        )
         
     
     def setup_initial_deque(self):
@@ -56,7 +61,10 @@ class directory_index_generator(object):
         
         return
     
-    def process_all_directories(self):
+    def process_all_directories(
+            self,
+            make_html=False,
+        ):
         """ Process all directories in the dir deque
         """
         
@@ -65,28 +73,34 @@ class directory_index_generator(object):
         while len(self.dir_deque) > 0:
             cur_dir = self.dir_deque.popleft()
             
-            self.process_directory(cur_dir)
+            self.process_directory(
+                cur_dir,
+                make_html=make_html,
+            )
         
         return
     
-    def process_directory(self, directory):
+    def process_directory(
+            self, directory,
+            make_html=False,
+        ):
         """ Process directory to determine all files and subdirectories
         """
         
         print(f'Dir: {directory}')
         
-        # Trim the base path to generate heirarchy
+        # Trim the base path to generate hierarchy and the directory name
         trim_prefix = self.base_yaml_path
         if trim_prefix[-1] != os.sep:
             trim_prefix = trim_prefix + os.sep
         
         trimmed_path = directory.replace(trim_prefix, '')
         
-        file_heirarchy = trimmed_path.split(os.sep)
+        file_hierarchy = trimmed_path.split(os.sep)
         
-        dir_name = file_heirarchy[-1]
+        dir_name = file_hierarchy[-1]
         
-        file_heirarchy = file_heirarchy[:-1]
+        file_hierarchy = file_hierarchy[:-1]
         
         # Scan the current directory for all contents
         dir_scan = os.scandir(directory)
@@ -97,7 +111,8 @@ class directory_index_generator(object):
         
         # Go through each content item
         for cur_item in dir_scan:
-            if cur_item.name == 'index.yaml':
+            # Ignore index files and Mac metadata files
+            if cur_item.name == 'index.yaml' or cur_item.name.startswith('._'):
                 continue
             
             if cur_item.is_dir():
@@ -121,8 +136,10 @@ class directory_index_generator(object):
         if len(cont_files) > 0:
             for cur_file in cont_files:
                 with open(f'{directory}/{cur_file}', 'r') as yaml_file:
-                    cur_yaml = yaml.load(yaml_file,
-                                         Loader=yaml.FullLoader)
+                    cur_yaml = yaml.load(
+                        yaml_file,
+                        Loader=yaml.FullLoader,
+                    )
                 
                 cur_file_name = cur_yaml['general']['name']
                 print(f'* {cur_file_name}')
@@ -132,14 +149,22 @@ class directory_index_generator(object):
                 yaml_file.close()
         
         
-        # Create index yaml for the directory
+        # Create an out dict that will be inserted into yaml file by Jinja
         out_dict = {}
         
         # Add general info
-        out_dict['general'] = {'name': dir_name,
-                               'file_heirarchy': file_heirarchy,
-                              }
+        if "{" in dir_name:
+            dir_name = '\'' + dir_name + '\''
         
+        out_dict['general'] = {
+            'name': dir_name,
+            'file_hierarchy': file_hierarchy,
+            'template': trimmed_path,
+            'releases': [self.release],
+            'environments': 'DR',
+        }
+        
+        # List of sub directories
         if len(cont_dirs) > 0:
             directories_dict = {}
             
@@ -152,18 +177,72 @@ class directory_index_generator(object):
             files_dict = {}
             
             for (file, file_name) in zip(cont_files, cont_file_names):
-                files_dict[file_name] = {'name': file_name,
-                                         'file': file.replace('.yaml', '')}
+                files_dict[file] = {
+                    'name': file_name,
+                    'file': file.replace('.yaml', ''),
+                }
             
             out_dict['file_species'] = files_dict
         
+        # Generate yaml file for the index
         out_yaml = f'{directory}/index.yaml'
-        with open(out_yaml, 'w') as out_yaml_file:
-            yaml.dump(out_dict, out_yaml_file)
         
+        if os.path.exists(out_yaml):
+            template_filler_dict = {}
+            template_filler_dict['general'] = {    
+                'manual_or_sw_generation': "replace me - with either manual or sw to indicate if this directory is made manually or generated by software",
+                'releases': [self.release],
+            }
+            
+            # Read in existing yaml file
+            with open(out_yaml, 'r') as file:
+                yaml_content = yaml.load(file, Loader=yaml.FullLoader)
+            
+            # Update list of sub directories and files
+            if len(cont_dirs) > 0:
+                yaml_content['directories'] = directories_dict
+            else:
+                # Remove directories that may have existed before but not anymore
+                yaml_content.pop('directories', None)
+            
+            if len(cont_files) > 0:
+                yaml_content['file_species'] = files_dict
+            else:
+                # Remove file species that may have existed before but not anymore
+                yaml_content.pop('file_species', None)
+            
+            # Update general
+            # Add missing keys
+            for key in template_filler_dict['general']:
+                if key not in yaml_content['general']:
+                    yaml_content['general'][key] = template_filler_dict['general'][key]
+            
+            # Update releases list
+            if self.release not in yaml_content['general']['releases']:
+                yaml_content['general']['releases'] = [self.release] + yaml_content['general']['releases']
+            
+            if isinstance(yaml_content['general']['releases'], dict):
+                yaml_content['general']['releases'] = dict(
+                    sorted(
+                        yaml_content['general']['releases'].items(), reverse=True,
+                    )
+                )
+            
+            # Write out updated yaml file
+            with open(out_yaml, 'w') as out_yaml_file:
+                yaml.dump(yaml_content, out_yaml_file)
+            
+        else:
+            self.template = self.environment.get_template('index.yaml')
+            yaml_out = yaml.load(
+                self.template.render(out_dict), Loader=yaml.FullLoader
+            )
+            with open(out_yaml, 'w') as out_yaml_file:
+                yaml.dump(yaml_out, out_yaml_file)
         
-        # Create an html file from the index yaml
-        self.create_html_output(out_yaml, trimmed_path)
+        if make_html:
+            # Create an html file from the index yaml
+            self.create_html_output(out_yaml, trimmed_path)
             
         return
